+++ hn_articles.csv
@@ -1,0 +1,30 @@
+title,url,domain,hn_id,comments_url,type,content_length,reported_comment_count,actual_comment_count,has_content,has_comments
+Gemini CLI,https://blog.google/technology/developers/introducing-gemini-cli-open-source-ai-agent/,blog.google,44376919,https://news.ycombinator.com/item?id=44376919,article,6551,764,764,True,True
+U.S. bombs Iranian nuclear sites,https://www.bbc.co.uk/news/live/ckg3rzj8emjt,bbc.co.uk,44341639,https://news.ycombinator.com/item?id=44341639,article,103,3832,3044,True,True
+Mechanical Watch: Exploded View,https://fellerts.no/projects/epoch.html,fellerts.no,44347425,https://news.ycombinator.com/item?id=44347425,article,22272,134,134,True,True
+Show HN: I'm an airline pilot – I built interactive graphs/globes of my flights,https://jameshard.ing/pilot,jameshard.ing,44396518,https://news.ycombinator.com/item?id=44396518,article,3160,157,152,True,True
+Writing toy software is a joy,https://blog.jsbarretto.com/post/software-is-joy,jsbarretto.com,44367084,https://news.ycombinator.com/item?id=44367084,article,13388,310,310,True,True
+"uv: An extremely fast Python package and project manager, written in Rust",https://github.com/astral-sh/uv,github.com,44357411,https://news.ycombinator.com/item?id=44357411,article,6730,334,334,True,True
+"OpenAI charges by the minute, so speed up your audio",https://george.mand.is/2025/06/openai-charges-by-the-minute-so-make-the-minutes-shorter/,mand.is,44376989,https://news.ycombinator.com/item?id=44376989,article,11514,221,218,True,True
+A new pyramid-like shape always lands the same side up,https://www.quantamagazine.org/a-new-pyramid-like-shape-always-lands-the-same-side-up-20250625/,quantamagazine.org,44381297,https://news.ycombinator.com/item?id=44381297,article,12716,158,150,True,True
+A new PNG spec,https://www.programmax.net/articles/png-is-back/,programmax.net,44365754,https://news.ycombinator.com/item?id=44365754,article,3358,592,587,True,True
+Fun with uv and PEP 723,https://www.cottongeeks.com/articles/2025-06-24-fun-with-uv-and-pep-723,cottongeeks.com,44369388,https://news.ycombinator.com/item?id=44369388,article,4410,224,224,True,True
+Backyard Coffee and Jazz in Kyoto,https://thedeletedscenes.substack.com/p/backyard-coffee-and-jazz-in-kyoto,substack.com,44356248,https://news.ycombinator.com/item?id=44356248,article,5856,265,265,True,True
+Vera C. Rubin Observatory first images,https://rubinobservatory.org/news/rubin-first-look/cosmic-treasure-chest,rubinobservatory.org,44356890,https://news.ycombinator.com/item?id=44356890,article,6926,165,165,True,True
+"Git Notes: Git's coolest, most unloved­ feature (2022)",https://tylercipriani.com/blog/2022/11/19/git-notes-gits-coolest-most-unloved-feature/,tylercipriani.com,44345334,https://news.ycombinator.com/item?id=44345334,article,5073,145,145,True,True
+Man 'refused entry into US' as border control catch him with bald JD Vance meme,https://www.dublinlive.ie/news/world-news/man-refused-entry-us-border-31925059,dublinlive.ie,44369140,https://news.ycombinator.com/item?id=44369140,article,5608,542,519,True,True
+Thnickels,https://thick-coins.net/?_bhlid=8a5736885893b7837e681aa73f890b9805a4673e,thick-coins.net,44372424,https://news.ycombinator.com/item?id=44372424,article,3193,124,124,True,True
+How I use my terminal,https://jyn.dev/how-i-use-my-terminal/,jyn.dev,44356646,https://news.ycombinator.com/item?id=44356646,article,12242,299,299,True,True
+Define policy forbidding use of AI code generators,https://github.com/qemu/qemu/commit/3d40db0efc22520fa6c399cf73960dced423b048,github.com,44382752,https://news.ycombinator.com/item?id=44382752,article,3982,387,387,True,True
+I wrote my PhD Thesis in Typst,https://fransskarman.com/phd_thesis_in_typst.html,fransskarman.com,44350322,https://news.ycombinator.com/item?id=44350322,article,13562,328,328,True,True
+-2000 Lines of code (2004),https://www.folklore.org/Negative_2000_Lines_Of_Code.html,folklore.org,44381252,https://news.ycombinator.com/item?id=44381252,article,3110,237,200,True,True
+AlphaGenome: AI for better understanding the genome,https://deepmind.google/discover/blog/alphagenome-ai-for-better-understanding-the-genome/,deepmind.google,44387659,https://news.ycombinator.com/item?id=44387659,article,12920,170,101,True,True
+What Problems to Solve (1966),http://genius.cat-v.org/richard-feynman/writtings/letters/problems,cat-v.org,44379606,https://news.ycombinator.com/item?id=44379606,article,3617,62,53,True,True
+Writing a basic Linux device driver when you know nothing about Linux drivers,https://crescentro.se/posts/writing-drivers/,crescentro.se,44345681,https://news.ycombinator.com/item?id=44345681,article,19490,67,67,True,True
+Microsoft Edit,https://github.com/microsoft/edit,github.com,44372380,https://news.ycombinator.com/item?id=44372380,article,2441,319,319,True,True
+"Starship: A minimal, fast, and customizable prompt for any shell",https://starship.rs/,starship.rs,44364874,https://news.ycombinator.com/item?id=44364874,article,108,211,211,True,True
+PlasticList – Plastic Levels in Foods,https://www.plasticlist.org/,plasticlist.org,44366548,https://news.ycombinator.com/item?id=44366548,article,0,189,189,False,True
+TPU Deep Dive,https://henryhmko.github.io/posts/tpu/tpu.html,github.io,44342977,https://news.ycombinator.com/item?id=44342977,article,22522,84,84,True,True
+Klein Bottle Amazon Brand Hijacking (2021),https://www.kleinbottle.com/Amazon_Brand_Hijacking.html,kleinbottle.com,44349525,https://news.ycombinator.com/item?id=44349525,article,8099,283,283,True,True
+LibRedirect – Redirects popular sites to alternative privacy-friendly frontends,https://libredirect.github.io/,github.io,44344246,https://news.ycombinator.com/item?id=44344246,article,1702,103,100,True,True
+"Games run faster on SteamOS than Windows 11, Ars testing finds",https://arstechnica.com/gaming/2025/06/games-run-faster-on-steamos-than-windows-11-ars-testing-finds/,arstechnica.com,44381144,https://news.ycombinator.com/item?id=44381144,article,7261,273,273,True,True
