--- podcast_cost_analysis.py
@@ -1,194 +1,0 @@
-#!/usr/bin/env python3
-"""
-Comprehensive Cost Optimization and Podcast Strategy Analysis
-"""
-
-from dynamodb_manager import DynamoDBManager
-from dotenv import load_dotenv
-import json
-
-load_dotenv()
-
-def analyze_podcast_cost_optimization():
-    """Comprehensive analysis of cost optimization for podcast generation."""
-    
-    print("🎙️ PODCAST GENERATION COST OPTIMIZATION ANALYSIS")
-    print("=" * 70)
-    
-    db = DynamoDBManager()
-    stats = db.get_stats()
-    
-    # Current state analysis
-    print("📊 CURRENT DATA STATE:")
-    print(f"   Articles: {stats['total_articles']}")
-    print(f"   Comments: {stats['total_comments']}")
-    print(f"   Comments per article: {stats['total_comments'] / max(stats['total_articles'], 1):.1f}")
-    
-    # Estimated storage costs
-    sample_article = db.get_articles(limit=1)[0] if db.get_articles(limit=1) else None
-    if sample_article:
-        comments = db.get_article_comments(sample_article['hn_id'])
-        if comments:
-            comment_size = len(json.dumps(comments[0]).encode('utf-8'))
-            total_comment_storage = comment_size * stats['total_comments']
-            
-            print(f"\n💾 STORAGE ANALYSIS:")
-            print(f"   Avg comment size: {comment_size} bytes")
-            print(f"   Total comment storage: {total_comment_storage / 1024:.1f} KB")
-            print(f"   Monthly storage cost: ${(total_comment_storage / (1024**3)) * 0.25:.6f}")
-    
-    print(f"\n🎯 PODCAST-OPTIMIZED STRATEGY:")
-    
-    # Strategy 1: Quality-based filtering
-    print(f"\n1️⃣ QUALITY-BASED COMMENT FILTERING:")
-    print(f"   Current: Store ALL comments (2,057 total)")
-    print(f"   Optimized: Store only high-quality comments")
-    print(f"   Criteria:")
-    print(f"     • Comment score ≥ 3 upvotes")
-    print(f"     • Content length ≥ 100 characters") 
-    print(f"     • Part of discussion thread (≥3 replies)")
-    print(f"     • No pure code/links")
-    print(f"   Expected reduction: ~70% fewer comments stored")
-    print(f"   Storage savings: ~70% reduction in comment storage costs")
-    print(f"   Quality improvement: Focus on engaging, discussion-worthy content")
-    
-    # Strategy 2: Article selection optimization
-    print(f"\n2️⃣ ARTICLE SELECTION OPTIMIZATION:")
-    print(f"   Current: Store articles with any engagement")
-    print(f"   Optimized: Store only podcast-worthy articles")
-    print(f"   Criteria:")
-    print(f"     • Article score ≥ 50 upvotes")
-    print(f"     • Comment count ≥ 10 comments")
-    print(f"     • Discussion-oriented (not job posts)")
-    print(f"   Expected reduction: ~60% fewer articles stored")
-    print(f"   Focus: High-engagement, discussion-rich content")
-    
-    # Strategy 3: Podcast generation costs
-    print(f"\n3️⃣ AI PODCAST GENERATION COSTS:")
-    print(f"   Model: GPT-4o-mini (cost-effective)")
-    print(f"   Input: ~1,500 tokens per article (comments + context)")
-    print(f"   Output: ~500 tokens per episode")
-    print(f"   Cost per episode: ~$0.001 (very affordable)")
-    print(f"   Daily cost (5 episodes): ~$0.005")
-    print(f"   Monthly cost: ~$0.15")
-    print(f"   Annual cost: ~$1.80")
-    
-    # Strategy 4: Read/Write optimization
-    print(f"\n4️⃣ READ/WRITE OPTIMIZATION:")
-    print(f"   Current pattern: Multiple individual comment reads")
-    print(f"   Optimized pattern: Batch processing and caching")
-    print(f"   Optimizations:")
-    print(f"     • Batch comment retrieval by article")
-    print(f"     • Cache podcast scripts to avoid regeneration")
-    print(f"     • Use DynamoDB query patterns efficiently")
-    print(f"     • Implement smart comment scoring system")
-    
-    # Total cost analysis
-    print(f"\n💰 TOTAL COST PROJECTION (OPTIMIZED):")
-    print(f"   DynamoDB Storage: ~$0.001/month")
-    print(f"   DynamoDB Read/Write: ~$0.01/month") 
-    print(f"   OpenAI API: ~$0.15/month")
-    print(f"   TOTAL: ~$0.16/month (~$1.92/year)")
-    print(f"   Cost per podcast episode: ~$0.032")
-    
-    print(f"\n🚀 IMPLEMENTATION RECOMMENDATIONS:")
-    
-    print(f"\n📝 IMMEDIATE ACTIONS:")
-    print(f"   1. Implement podcast-optimized scraper")
-    print(f"   2. Add comment scoring to database schema")
-    print(f"   3. Create discussion thread detection")
-    print(f"   4. Set up OpenAI integration with proper error handling")
-    print(f"   5. Implement podcast episode caching")
-    
-    print(f"\n📈 ADVANCED OPTIMIZATIONS:")
-    print(f"   1. Machine learning comment quality scoring")
-    print(f"   2. User engagement prediction")
-    print(f"   3. Topic clustering for episode themes")
-    print(f"   4. Automated quality metrics for podcast content")
-    print(f"   5. A/B testing for different podcast formats")
-    
-    print(f"\n🎧 PODCAST QUALITY FEATURES:")
-    print(f"   1. Extract discussion themes automatically")
-    print(f"   2. Identify contrasting viewpoints")
-    print(f"   3. Highlight expert insights (high-karma users)")
-    print(f"   4. Include relevant context and background")
-    print(f"   5. Generate engaging transitions between topics")
-    
-    return {
-        'current_comments': stats['total_comments'],
-        'optimized_comments_estimate': int(stats['total_comments'] * 0.3),
-        'storage_savings_percent': 70,
-        'monthly_cost_estimate': 0.16,
-        'cost_per_episode': 0.032
-    }
-
-def generate_implementation_roadmap():
-    """Generate a roadmap for implementing the podcast system."""
-    
-    print(f"\n🗺️ IMPLEMENTATION ROADMAP")
-    print("=" * 50)
-    
-    phases = [
-        {
-            'phase': 'Phase 1: Foundation (Week 1-2)',
-            'tasks': [
-                'Fix OpenAI library compatibility',
-                'Implement comment scoring in scraper',
-                'Add discussion thread detection',
-                'Create podcast episode storage schema',
-                'Basic AI prompt engineering'
-            ]
-        },
-        {
-            'phase': 'Phase 2: Quality Optimization (Week 3-4)', 
-            'tasks': [
-                'Implement quality-based comment filtering',
-                'Add article selection criteria',
-                'Create engagement scoring system',
-                'Optimize database queries',
-                'Add error handling and retry logic'
-            ]
-        },
-        {
-            'phase': 'Phase 3: Podcast Generation (Week 5-6)',
-            'tasks': [
-                'Perfect AI prompt for engaging narratives',
-                'Implement multi-topic episode generation',
-                'Add podcast metadata and indexing',
-                'Create episode quality metrics',
-                'Add automated publishing pipeline'
-            ]
-        },
-        {
-            'phase': 'Phase 4: Advanced Features (Week 7-8)',
-            'tasks': [
-                'ML-based comment quality prediction',
-                'Topic clustering and theme detection',
-                'Personalized episode recommendations',
-                'Real-time discussion monitoring',
-                'Analytics and performance tracking'
-            ]
-        }
-    ]
-    
-    for phase_info in phases:
-        print(f"\n{phase_info['phase']}")
-        print("-" * 40)
-        for i, task in enumerate(phase_info['tasks'], 1):
-            print(f"   {i}. {task}")
-    
-    print(f"\n🎯 SUCCESS METRICS:")
-    print(f"   • Episode generation time < 30 seconds")
-    print(f"   • Comment relevance score > 80%")
-    print(f"   • Storage cost < $2/month")
-    print(f"   • Episode engagement > 5 min listen time")
-    print(f"   • User retention > 70%")
-
-if __name__ == "__main__":
-    analysis = analyze_podcast_cost_optimization()
-    generate_implementation_roadmap()
-    
-    print(f"\n✅ ANALYSIS COMPLETE")
-    print(f"Expected monthly savings: {analysis['storage_savings_percent']}%")
-    print(f"Estimated system cost: ${analysis['monthly_cost_estimate']:.2f}/month")
-    print(f"Ready for podcast implementation! 🎙️")
